refactor(util, game) : 숫자 유효성검사 패키지 이동
유틸패키지에서 게임패키지 내부로 이동
